+++ CHANGELOG.md
@@ -14,9 +14,11 @@
  - Fix #168: Fix spelling in russian language (EvgenyOrekhov)
  - Fix #195: UserCreateService: check if we're from web before setting flash message (maxxer)
  - Enh: Improvements to the admin responsive design (wautvda)
+ - Enh: Add controller module class reference (TonisOrmisson)
  - Enh: Replace the deprecated InvalidParamException in ClassMapHelper (TonisOrmisson)
+ - Fix #242: Add POST filter for `admin/force-password-change` action (bscheshirwork)
  - Enh #249: Show message `email send if possible` any time on reset password request (bscheshirwork)
- 
+
 ## 1.1.4 - February 19, 2018
 - Enh: Check enableEmailConfirmation on registration (faenir)
 - Fix #154: Fix DateTime constructor with Unix timestamps (tonydspaniard)